requirements.txt
Flask==2.1.2
nltk==3.6.2
pyttsx3==2.90
pyautogui==0.9.53
Pillow==8.3.1  # For the PIL module (Pillow is the modern version of PIL)
pydub==0.25.1
numpy==1.21.1
sounddevice==0.4.2
soundfile==0.10.3.post1
soundcard==0.4.4
